fix small bugs

+++ learning_based/sample_generator.py
@@ -32,7 +32,7 @@ class generator(object):
     def QAM_N_const(self, n=None):
         if n==None:
             n = self.mod_n
-        constellation = np.linspace(-np.sqrt(n)+1, np.sqrt(n)-1, np.sqrt(n))
+        constellation = np.linspace(int(-np.sqrt(n)+1), int(np.sqrt(n)-1), int(np.sqrt(n)))
         alpha = np.sqrt((constellation ** 2).mean())
         constellation /= (alpha * np.sqrt(2))
         constellation = tf.Variable(constellation, trainable=False, dtype=tf.float32)
